Widen HIR expression ids

+++ crates/tune_hir/src/lib.rs
@@ -9,7 +9,7 @@ pub mod shape;
 pub struct HirId(pub u32);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct ExprId(pub u32);
+pub struct ExprId(pub u64);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum MemberKind {
